+++ modules/react-focus-navigation/src/__tests__/useEventHandler.spec.lua
@@ -23,7 +23,7 @@ local Collections = require(Packages.Dev.Collections)
 local Object = Collections.Object
 
 local FocusNavigationContext = require(script.Parent.Parent.FocusNavigationContext)
-local useEventHandlerMap = require(script.Parent.Parent.useEventHandlerMap)
+local useEventHandler = require(script.Parent.Parent.useEventHandler)
 
 local focusNavigationService
 beforeEach(function()
@@ -37,7 +37,7 @@ afterEach(function()
 end)
 
 local function SimpleButton(props)
-	local ref = useEventHandlerMap(props.handlerMap, props.innerRef)
+	local ref = useEventHandler(props.eventName, props.eventHandler, props.eventPhase, props.innerRef)
 	return React.createElement("TextButton", {
 		Text = props.text,
 		ref = ref,
@@ -45,7 +45,7 @@ local function SimpleButton(props)
 end
 
 local function SimpleLabel(props)
-	local ref = useEventHandlerMap(props.handlerMap, props.innerRef)
+	local ref = useEventHandler(props.eventName, props.eventHandler, props.eventPhase, props.innerRef)
 	return React.createElement("TextLabel", {
 		Text = props.text,
 		ref = ref,
@@ -66,9 +66,8 @@ it("should have no effect if no context is provided", function()
 	local handler, handlerFn = jest.fn()
 	-- Use regular testing library `render` instead of the wrapper
 	local result = render(React.createElement(SimpleButton, {
-		handlerMap = {
-			focus = { handler = handlerFn },
-		},
+		eventName = "onXButton",
+		eventHandler = handlerFn,
 		text = "Confirm",
 	}))
 
@@ -86,34 +85,44 @@ it("should have no effect if no context is provided", function()
 end)
 
 it("should receive focus and blur events when registered", function()
-	local focusHandler, focusHandlerFn = jest.fn()
-	local blurHandler, blurHandlerFn = jest.fn()
-	local result = renderWithFocusNav(React.createElement(SimpleButton, {
-		handlerMap = {
-			focus = { handler = focusHandlerFn },
-			blur = { handler = blurHandlerFn },
-		},
-		text = "Confirm",
+	local leftFocusHandler, focusHandlerFn = jest.fn()
+	local rightBlurHandler, blurHandlerFn = jest.fn()
+	local result = renderWithFocusNav(React.createElement(React.Fragment, nil, {
+		LeftButton = React.createElement(SimpleButton, {
+			eventName = "focus",
+			eventHandler = focusHandlerFn,
+			text = "Confirm",
+		}),
+		RightButton = React.createElement(SimpleButton, {
+			eventName = "blur",
+			eventHandler = blurHandlerFn,
+			text = "Cancel",
+		}),
 	}))
 
-	local instance = result.getByText("Confirm")
-	focusNavigationService:focusGuiObject(instance, false)
+	local leftButton = result.getByText("Confirm")
+	focusNavigationService:focusGuiObject(leftButton, false)
 
-	expect(focusHandler).toHaveBeenCalledTimes(1)
-	expect(blurHandler).toHaveBeenCalledTimes(0)
+	expect(leftFocusHandler).toHaveBeenCalledTimes(1)
+	expect(rightBlurHandler).toHaveBeenCalledTimes(0)
 
-	focusNavigationService:focusGuiObject(nil, false)
+	local rightButton = result.getByText("Cancel")
+	focusNavigationService:focusGuiObject(rightButton, false)
 
-	expect(focusHandler).toHaveBeenCalledTimes(1)
-	expect(blurHandler).toHaveBeenCalledTimes(1)
+	expect(leftFocusHandler).toHaveBeenCalledTimes(1)
+	expect(rightBlurHandler).toHaveBeenCalledTimes(0)
+
+	focusNavigationService:focusGuiObject(leftButton, false)
+
+	expect(leftFocusHandler).toHaveBeenCalledTimes(2)
+	expect(rightBlurHandler).toHaveBeenCalledTimes(1)
 end)
 
 it("should deregister an event map on unmount", function()
 	local handler, handlerFn = jest.fn()
 	local result = renderWithFocusNav(React.createElement(SimpleButton, {
-		handlerMap = {
-			onXButton = { handler = handlerFn },
-		},
+		eventName = "onXButton",
+		eventHandler = handlerFn,
 		text = "Confirm",
 	}))
 
@@ -150,13 +159,47 @@ it("should deregister an event map on unmount", function()
 	expect(handler).toHaveBeenCalledTimes(2)
 end)
 
+it("should deregister old and register new handlers when they change", function()
+	local handler, handlerFn = jest.fn()
+	local result = renderWithFocusNav(React.createElement(SimpleButton, {
+		eventName = "onXButton",
+		eventHandler = handlerFn,
+		text = "Confirm",
+	}))
+
+	local instance = result.getByText("Confirm")
+	focusNavigationService:registerEventMap(instance, {
+		[Enum.KeyCode.ButtonX] = "onXButton",
+	})
+	focusNavigationService:focusGuiObject(instance, false)
+
+	local gamepad = GamePad.new()
+	act(function()
+		gamepad:hitButton(Enum.KeyCode.ButtonX)
+	end)
+	-- press + release
+	expect(handler).toHaveBeenCalledTimes(2)
+
+	local newHandler, newHandlerFn = jest.fn()
+	result.rerender(React.createElement(SimpleButton, {
+		eventName = "onXButton",
+		eventHandler = newHandlerFn,
+		text = "Confirm",
+	}))
+
+	act(function()
+		gamepad:hitButton(Enum.KeyCode.ButtonX)
+	end)
+	expect(handler).toHaveBeenCalledTimes(2)
+	expect(newHandler).toHaveBeenCalledTimes(2)
+end)
+
 it("should update the event map if the value changes after it updates", function()
 	local handlerA, handlerAFn = jest.fn()
 	local handlerB, handlerBFn = jest.fn()
 	local result = renderWithFocusNav(React.createElement(SimpleButton, {
-		handlerMap = {
-			onXButton = { handler = handlerAFn },
-		},
+		eventName = "onXButton",
+		eventHandler = handlerAFn,
 		text = "Confirm",
 	}))
 
@@ -174,9 +217,8 @@ it("should update the event map if the value changes after it updates", function
 	expect(handlerB).toHaveBeenCalledTimes(0)
 
 	result.rerender(React.createElement(SimpleButton, {
-		handlerMap = {
-			onXButton = { handler = handlerBFn },
-		},
+		eventName = "onXButton",
+		eventHandler = handlerBFn,
 		text = "Confirm",
 	}))
 	act(function()
@@ -190,15 +232,15 @@ end)
 it("should call the correct handler if the eventMap updates", function()
 	local onXHandler, onXHandlerFn = jest.fn()
 	local onYHandler, onYHandlerFn = jest.fn()
-	local result = renderWithFocusNav(React.createElement(SimpleButton, {
-		handlerMap = {
-			onXButton = { handler = onXHandlerFn },
-			onYButton = { handler = onYHandlerFn },
-		},
-		text = "Confirm",
-	}))
+	local function TwoHandlersButton(_props)
+		local innerRef = useEventHandler("onXButton", onXHandlerFn)
+		local ref = useEventHandler("onYButton", onYHandlerFn, nil, innerRef)
+		return React.createElement("TextButton", { ref = ref, Text = "Hello" })
+	end
 
-	local instance = result.getByText("Confirm")
+	local result = renderWithFocusNav(React.createElement(TwoHandlersButton))
+
+	local instance = result.getByText("Hello")
 	focusNavigationService:registerEventMap(instance, {
 		[Enum.KeyCode.ButtonX] = "onXButton",
 	})
@@ -232,7 +274,7 @@ end)
 -- called after some time? Maybe in dev mode?
 it("should not do anything if the returned ref is not used", function()
 	local function IgnoresRef(props)
-		local _ref = useEventHandlerMap(props.handlerMap)
+		local _ref = useEventHandler(props.eventName, props.eventHandler)
 		return React.createElement("TextButton", {
 			Text = props.text,
 		})
@@ -240,9 +282,8 @@ it("should not do anything if the returned ref is not used", function()
 
 	local handler, handlerFn = jest.fn()
 	local result = renderWithFocusNav(React.createElement(IgnoresRef, {
-		handlerMap = {
-			onXButton = { handler = handlerFn },
-		},
+		eventName = "onXButton",
+		eventHandler = handlerFn,
 		text = "Confirm",
 	}))
 
@@ -260,8 +301,10 @@ end)
 
 it("should update a function ref when the underlying value changes", function()
 	local refSpy, refSpyFn = jest.fn()
+	local mockHandler = function() end
 	local result = renderWithFocusNav(React.createElement(SimpleButton, {
-		handlerMap = {},
+		eventName = "someEvent",
+		eventHandler = mockHandler,
 		innerRef = refSpyFn,
 		text = "Confirm",
 	}))
@@ -272,7 +315,8 @@ it("should update a function ref when the underlying value changes", function()
 	expect(refSpy).toHaveBeenCalledWith(instance)
 
 	result.rerender(React.createElement(SimpleLabel, {
-		handlerMap = {},
+		eventName = "someEvent",
+		eventHandler = mockHandler,
 		innerRef = refSpyFn,
 		text = "Confirm",
 	}))
@@ -286,8 +330,10 @@ end)
 
 it("should update an object ref when the underlying value changes", function()
 	local ref = React.createRef()
+	local mockHandler = function() end
 	local result = renderWithFocusNav(React.createElement(SimpleButton, {
-		handlerMap = {},
+		eventName = "someEvent",
+		eventHandler = mockHandler,
 		innerRef = ref,
 		text = "Confirm",
 	}))
@@ -295,11 +341,49 @@ it("should update an object ref when the underlying value changes", function()
 	expect(ref.current).toEqual(instance)
 
 	result.rerender(React.createElement(SimpleLabel, {
-		handlerMap = {},
+		eventName = "someEvent",
+		eventHandler = mockHandler,
 		innerRef = ref,
 		text = "Confirm",
 	}))
 
 	local newInstance = result.getByText("Confirm")
 	expect(ref.current).toEqual(newInstance)
+end)
+
+it("binds to the correct phase", function()
+	local captureFocusHandler, captureFocusHandlerFn = jest.fn()
+	local bubbleFocusHandler, bubbleFocusHandlerFn = jest.fn()
+	local result = renderWithFocusNav(React.createElement(React.Fragment, nil, {
+		LeftButton = React.createElement(SimpleButton, {
+			eventName = "focus",
+			eventHandler = captureFocusHandlerFn,
+			eventPhase = "Capture" :: FocusNavigation.EventPhase,
+			text = "Confirm",
+		}),
+		RightButton = React.createElement(SimpleButton, {
+			eventName = "focus",
+			eventHandler = bubbleFocusHandlerFn,
+			eventPhase = "Bubble" :: FocusNavigation.EventPhase,
+			text = "Cancel",
+		}),
+	}))
+
+	local leftButton = result.getByText("Confirm")
+	focusNavigationService:focusGuiObject(leftButton, false)
+
+	expect(captureFocusHandler).toHaveBeenCalledTimes(1)
+	expect(captureFocusHandler).toHaveBeenCalledWith(expect.objectContaining({
+		phase = "Capture",
+	}))
+	expect(bubbleFocusHandler).toHaveBeenCalledTimes(0)
+
+	local rightButton = result.getByText("Cancel")
+	focusNavigationService:focusGuiObject(rightButton, false)
+
+	expect(captureFocusHandler).toHaveBeenCalledTimes(1)
+	expect(bubbleFocusHandler).toHaveBeenCalledTimes(1)
+	expect(bubbleFocusHandler).toHaveBeenCalledWith(expect.objectContaining({
+		phase = "Bubble",
+	}))
 end)